refactor(frontend): split session hydration stages
