+++ packages/frontend/src/features/agent-studio-build-tools/use-agent-studio-build-tools-worktree-snapshot.test.tsx
@@ -179,6 +179,35 @@ describe("useAgentStudioBuildToolsWorktreeSnapshot", () => {
     }
   });
 
+  test("keeps diff and status hydration scoped to the direct build worktree", async () => {
+    const harness = createHookHarness(
+      createBaseArgs({
+        session: {
+          role: "build",
+          status: "running",
+          workingDirectory: "/repo/.worktrees/task-24",
+          hasActiveSession: true,
+        },
+        worktreeRecoverySignal: 5,
+      }),
+    );
+
+    try {
+      await harness.mount();
+
+      expect(taskWorktreeGetMock).not.toHaveBeenCalled();
+      expect(harness.getLatest().diffData.worktreePath).toBe("/repo/.worktrees/task-24");
+      expect(useAgentStudioDiffDataMock.mock.calls.at(-1)?.[0]).toMatchObject({
+        repoPath: "/repo",
+        worktreePath: "/repo/.worktrees/task-24",
+        worktreeResolutionTaskId: null,
+        shouldBlockDiffLoading: false,
+      });
+    } finally {
+      await harness.unmount();
+    }
+  });
+
   test("queries the canonical task worktree when no direct worktree exists", async () => {
     const harness = createHookHarness(createBaseArgs());
