+++ packages/frontend/src/state/operations/agent-orchestrator/lifecycle/load-sessions-stage-constants.ts
@@ -1,0 +1,5 @@
+import type { RepoPromptOverrides } from "@openducktor/contracts";
+
+export const INITIAL_SESSION_HISTORY_LIMIT = 600;
+export const SESSION_HISTORY_HYDRATION_CONCURRENCY = 3;
+export const EMPTY_PROMPT_OVERRIDES: RepoPromptOverrides = {};
